Cleaned libft to be used with CFLAGS

+++ ft_printf/ft_dprintf.c
@@ -6,16 +6,16 @@
 /*   By: rorollin <rorollin@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/07 15:48:08 by rorollin          #+#    #+#             */
-/*   Updated: 2025/02/12 20:36:56 by rorollin         ###   ########.fr       */
+/*   Updated: 2025/04/01 16:01:55 by rorollin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_dprintf(int fd, const char *str, ...)
+ssize_t	ft_dprintf(int fd, const char *str, ...)
 {
 	va_list	arguments;
-	size_t	bytes_written;
+	ssize_t	bytes_written;
 
 	bytes_written = 0;
 	if (str == NULL)
